requirements.txt
google-generativeai
pandas
numpy
python-dotenv
pyyaml
pydantic
tabulate
matplotlib
seaborn